+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added new function `as_forest_plot()` that converts a gtsummary table into a forest plot using `forestplot::forestplot()`.
 
+* Added new function `plot_cuminc()` to plot cumulative incidence estimates with both the at-risk table and the estimates printed below the table.
+
 * Updated Rmd files in project template to use user's custom `gtsummary` theme if it exists within `bstfun::theme_gtsummary_msk()`, matching the system login name to the `name` in `bstfun::theme_gtsummary_msk()`
 
 * Adding new function `use_bst_rstudio_prefs()`