+++ NEWS.md
@@ -1,5 +1,7 @@
 # bstfun (development version)
 
+* Deprecated `tbl_2way_summary()` in favor of `gtsummary::tbl_continous()`.
+
 # bstfun 0.2.5
 
 * Deprecated `bstfun::tbl_split()`, `bstfun::gts_add_p_footnote()`, and `bstfun::gtsummary_butcher()` in lieu of `gtsummary::tbl_split()`, `gtsummary::separate_p_footnote()`, and `gtsummary::tbl_butcher()`.